Moving displayReceivedFiles from server to shared utility.
